refactor(content): reclassify sequential-task-executor as logic
Per review feedback, sequential-task-executor is a pure in-process
concurrency primitive (PQueue wrapper) with no I/O. It fits better
under logic/ than adapters/.

Moved src/adapters/sequential-task-executor/ -> src/logic/sequential-task-executor/.
Importers in components.ts, types.ts, metrics.ts, the test mock, and
test/unit/ports/deployer.spec.ts updated via a single sed pass.

tsc clean, unit suite green (213 tests, no diff).
