+++ content/src/components.ts
@@ -27,7 +27,7 @@ import { createDeployer } from './ports/deployer'
 import { createFailedDeployments } from './ports/failedDeployments'
 import { createDatabaseComponent } from './ports/postgres'
 import { createProcessedSnapshotStorage } from './ports/processedSnapshotStorage'
-import { createSequentialTaskExecutor } from './adapters/sequential-task-executor'
+import { createSequentialTaskExecutor } from './logic/sequential-task-executor'
 import { createSnapshotGenerator } from './ports/snapshotGenerator'
 import { createSnapshotStorage } from './ports/snapshotStorage'
 import { createSynchronizationState } from './ports/synchronizationState'
